WIP remaining 2to3 changes, probably quite wrong

+++ src/foolscap/test/common.py
@@ -72,7 +72,7 @@ class Loopback:
 Digits = re.compile("\d*")
 MegaSchema1 = DictOf(str,
                      ListOf(TupleOf(SetOf(int, maxLength=10, mutable=True),
-                                    str, bool, int, long, float, None,
+                                    str, bool, int, int, float, None,
                                     UnicodeConstraint(),
                                     ByteStringConstraint(),
                                     Any(), NumberConstraint(),
